+++ shuup/core/api/__init__.py
@@ -10,7 +10,11 @@ from shuup.core.api.attribute import AttributeViewSet
 from shuup.core.api.contacts import ContactViewSet
 from shuup.core.api.manufacturer import ManufacturerViewSet
 from shuup.core.api.orders import OrderViewSet
-from shuup.core.api.products import ProductViewSet, ShopProductViewSet
+from shuup.core.api.product_media import ProductMediaViewSet
+from shuup.core.api.products import (
+    ProductAttributeViewSet, ProductPackageViewSet, ProductTypeViewSet,
+    ProductViewSet, ShopProductViewSet
+)
 from shuup.core.api.shipments import ShipmentViewSet
 from shuup.core.api.suppliers import SupplierViewSet
 from shuup.core.api.tax_class import TaxClassViewSet
@@ -29,6 +33,10 @@ def populate_core_api(router):
     router.register("shuup/contact", ContactViewSet)
     router.register("shuup/order", OrderViewSet)
     router.register("shuup/product", ProductViewSet)
+    router.register("shuup/product_attribute", ProductAttributeViewSet)
+    router.register("shuup/product_media", ProductMediaViewSet)
+    router.register("shuup/product_type", ProductTypeViewSet)
+    router.register("shuup/product_package", ProductPackageViewSet)
     router.register("shuup/shipment", ShipmentViewSet)
     router.register("shuup/shop", viewset_factory(Shop))
     router.register("shuup/shop_product", ShopProductViewSet)